Change license to MIT.
As the project is still young and there are yet any forks, I feel that it is perfectly acceptable to change the license at this time. Not only that, I am changing the license to the more permissive MIT (previously GPL).

My intentions for this application were for future users to be able to fork this code and make modifications needed to run in a more locked down environment (for example: to project the snippets being shared openly so you could feel safe sharing snippets within a more controlled system), and I feel that the GPL would have made that more difficult (if not impossible), so I have decided to change to the MIT license.